# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/hackathons": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all Hackathons",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Get all Hackathons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Hackathon"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Hackathon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Create a new Hackathon",
                "parameters": [
                    {
                        "description": "Hackathon object",
                        "name": "hackathon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HackathonCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created Hackathon",
                        "schema": {
                            "$ref": "#/definitions/models.Hackathon"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hackathons/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get hackathons created by the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Get hackathons created by the user",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of hackathons",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Hackathon"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hackathons/{hackathonId}/teams": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all teams for a specific hackathon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get all teams for a specific hackathon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hackathon ID",
                        "name": "hackathonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Team"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hackathons/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a single Hackathon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Get a single Hackathon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hackathon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved Hackathon",
                        "schema": {
                            "$ref": "#/definitions/models.Hackathon"
                        }
                    },
                    "404": {
                        "description": "Hackathon not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a Hackathon",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Update a Hackathon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hackathon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hackathon object",
                        "name": "hackathon",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HackathonCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated Hackathon",
                        "schema": {
                            "$ref": "#/definitions/models.Hackathon"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Hackathon not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a Hackathon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Delete a Hackathon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hackathon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted Hackathon",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/hackathons/{id}/register": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register for a Hackathon",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hackathons"
                ],
                "summary": "Register for a Hackathon",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hackathon ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully registered for Hackathon",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.File": {
            "type": "object",
            "properties": {
                "contentType": {
                    "description": "File content type",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hackathon_id": {
                    "description": "Foreign key for Hackathon",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastModified": {
                    "description": "Last modified date",
                    "type": "string"
                },
                "name": {
                    "description": "Filename",
                    "type": "string"
                },
                "size": {
                    "description": "File size in bytes",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "description": "URL to download the file from Google Cloud Storage",
                    "type": "string"
                },
                "userId": {
                    "description": "User ID",
                    "type": "string"
                }
            }
        },
        "models.Hackathon": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/models.User"
                },
                "created_by_id": {
                    "type": "integer"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "hackathon_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "location": {
                    "type": "string"
                },
                "max_participants": {
                    "type": "integer"
                },
                "max_participants_per_team": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nb_of_teams": {
                    "type": "integer"
                },
                "participations": {
                    "description": "Many-to-many relationship with User through Participation",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Participation"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Team"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.HackathonCreate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Paris"
                },
                "description": {
                    "type": "string",
                    "example": "Un événement pour les développeurs"
                },
                "end_date": {
                    "type": "string",
                    "example": "2021-01-02"
                },
                "location": {
                    "type": "string",
                    "example": "Paris"
                },
                "max_participants": {
                    "type": "integer",
                    "example": 100
                },
                "max_participants_per_team": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Hackathon de Paris"
                },
                "start_date": {
                    "type": "string",
                    "example": "2021-01-01"
                }
            }
        },
        "models.Participation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "hackathonID": {
                    "description": "Foreign key referencing Hackathon.ID",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isOrganizer": {
                    "description": "Indicates if the user is an organizer of the hackathon",
                    "type": "boolean"
                },
                "team_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "description": "Foreign key referencing User.ID",
                    "type": "integer"
                }
            }
        },
...
        "models.Team": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "evaluation": {
                    "description": "Relation avec Evaluation",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Evaluation"
                        }
                    ]
                },
                "evaluation_id": {
                    "description": "Ajout de l'EvaluationID",
                    "type": "integer"
                },
                "hackathon": {
                    "$ref": "#/definitions/models.Hackathon"
                },
                "hackathon_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "Team 1"
                },
                "nbOfMembers": {
                    "type": "integer"
                },
                "submission": {
                    "description": "Assuming there is a Submission model",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Submission"
                        }
                    ]
                },
                "submission_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "users": {
                    "description": "Has many Users",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}